Head Tracker Second Commit
In this phase, I managed to print in the console the number of times, the user has looked in various directions, either left or right. I have made use of a counter in the code so that I can get this type of data from my head tracker tool to analyze the user interactions with the design. The counter updates by one, every time the user changes his head directions to navigate, and due to this I can obtain such data for my study.